Renamed SnS folders
