+++ src/factiva/analytics/__init__.py
@@ -17,8 +17,8 @@ from .__version__ import __version__
 from .article_retrieval import ArticleRetrieval, UIArticle
 from .auth import UserKey, OAuthUser
 from .taxonomy import FactivaTaxonomy, FactivaTaxonomyCategories
-from .snapshot import SnapshotExplain, SnapshotExplainQuery, SnapshotExplainJobResponse, SnapshotExplainSamplesResponse
-from .snapshot import SnapshotTimeSeries, SnapshotTimeSeriesQuery, SnapshotTimeSeriesJobReponse
+from .snapshots import SnapshotExplain, SnapshotExplainQuery, SnapshotExplainJobResponse, SnapshotExplainSamplesResponse
+from .snapshots import SnapshotTimeSeries, SnapshotTimeSeriesQuery, SnapshotTimeSeriesJobReponse
 # from .common.tools import SnapshotFiles
 # from .common.log import factiva_logger, get_factiva_logger
 # from .streaming import StreamingInstance, Listener, Subscription
