+++ src/factiva/analytics/snapshots/base.py
