+++ src/factiva/analytics/snapshots/bulknews.py
