+++ src/factiva/analytics/streams/subscription.py
